[Refactor] (waitings): waiting list pocketbase realtime based
