+++ frontend/hooks/waits/useFetchWaitList.ts
@@ -1,6 +1,6 @@
 'use client'
-import GetCompanyInfo from "@/server-actions/waits/GetCompanyInfo"
-import GetUserWait from "@/server-actions/waits/GetUserWait"
+import client from "@/libs/pocketbase"
+import { GetCompanyInfo, GetUserWait } from "@/server-actions/waits/waits"
 import { UseFetchWaitListResult, UserWaitParams } from "@/types/waits/types"
 import { useParams } from "next/navigation"
 import { useEffect, useState } from "react"
@@ -17,6 +17,8 @@ const useFetchWaitList = (): UseFetchWaitListResult => {
     const [waitUserList, setWaitUserList] = useState<UserWaitParams[]>([])
     const [phoneNumber, setPhoneNumber] = useState<string>("010-")
     const [isFetching, setIsFetching] = useState<boolean>(false)
+    const [action, setAction] = useState<string>("")
+
     const params = useParams()
 
     useEffect(() => {
@@ -43,7 +45,7 @@ const useFetchWaitList = (): UseFetchWaitListResult => {
 
                 setRulesEnabled(data?.expand?.management_waits[0]?.rules_enabled)
                 setRulesContent(data?.expand?.management_waits[0]?.rules_content)
-                
+
                 setLimitPerson(data?.expand?.management_waits[0]?.limit_persons)
 
 
@@ -54,11 +56,40 @@ const useFetchWaitList = (): UseFetchWaitListResult => {
 
         fetchData()
 
-    }, [isFetching]);
+    }, [action == null]);
+
+    useEffect(() => {
+        const fetchManageData = async () => {
+            const response = await GetUserWait(manageId)
+            const userWaitData = response.expand?.user_waits ? response.expand?.user_waits : []
+            const trueAdmissionStatusList = userWaitData?.filter((item: UserWaitParams) => item.admission_status === false);
+            setUserWaitsNumber(trueAdmissionStatusList?.length ? trueAdmissionStatusList.length : 0)
+            setWaitUserList(userWaitData)
+        }
+
+        // manageId가 변경될 때마다 이전 구독을 취소하고 다시 구독
+        const userWaitSubscribe = client.collection('user_waits').subscribe('*', function (e) {
+            if (manageId.length >= 1) {
+                fetchManageData();
+            }
+        });
+
+        const managementWaitSubscribe = client.collection('management_waits').subscribe('*', function (e) {
+            if (manageId.length >= 1) {
+                fetchManageData();
+            }
+        })
+
+        // 컴포넌트가 언마운트되거나 manageId가 변경될 때 이전 구독 취소
+        return () => {
+            client.collection('user_waits').unsubscribe()
+            client.collection('management_waits').unsubscribe()
+        };
+    }, [manageId, action != null])
 
     return [manageId, companyId, companyName, waitTime, userWaitsNumber, rulesEnabled,
         rulesContent, limitPerson, waitUserList, phoneNumber,
-        setPhoneNumber, setIsFetching 
+        setPhoneNumber, action
     ]
 }
